test: move all test files to /tests directory
