--- class-components/src/components/__tests__/Header.test.tsx
@@ -1,14 +1,0 @@
-import { screen } from '@testing-library/react';
-
-import { renderWithRouter } from '../../utils';
-import { Header } from '..';
-
-describe('Header', () => {
-  it('should render correctly', () => {
-    renderWithRouter(<Header />);
-
-    expect(screen.getByText(/Star Wars planet finder app/i)).toBeInTheDocument();
-
-    expect(screen.getByRole('link')).toBeInTheDocument();
-  });
-});
