+++ class-components/src/tests/setup-tests.ts
@@ -1,0 +1,7 @@
+import '@testing-library/jest-dom';
+
+import { server } from './mocks/server';
+
+beforeAll(() => server.listen());
+afterEach(() => server.resetHandlers());
+afterAll(() => server.close());
